Make realistic board rendering the default
