+++ src/textures/create-three-texture-meshes.ts
@@ -1,10 +1,9 @@
 import type { PcbBoard } from "circuit-json"
 import * as THREE from "three"
 import { REALISTIC_BOARD_SURFACE_MATERIAL } from "../board-surface-textures"
-import type { RenderingMode } from "../contexts/RenderingModeContext"
 import { FAUX_BOARD_OPACITY } from "../geoms/constants"
-import { createBoardReliefTextures } from "../utils/create-board-relief-textures"
 import { configureObjectShadows } from "../utils/configure-object-shadows"
+import { createBoardReliefTextures } from "../utils/create-board-relief-textures"
 import { createBoardShadowReceiverPlane } from "../utils/create-board-shadow-receiver-plane"
 import { calculateOutlineBounds } from "../utils/outline-bounds"
 import type { CombinedBoardTextures } from "./index"
@@ -16,7 +15,6 @@ interface TexturePlaneConfig {
   usePolygonOffset?: boolean
   renderOrder?: number
   isFaux?: boolean
-  renderingMode?: RenderingMode
 }
 
 function createTexturePlane(
@@ -30,7 +28,6 @@ function createTexturePlane(
     usePolygonOffset = false,
     renderOrder = 0,
     isFaux = false,
-    renderingMode = "engineering",
   } = config
 
   if (!texture) return null
@@ -55,32 +52,28 @@ function createTexturePlane(
     opacity: isFaux ? FAUX_BOARD_OPACITY : 1.0,
   } satisfies THREE.MeshBasicMaterialParameters
 
-  const reliefTextures =
-    renderingMode === "realistic" ? createBoardReliefTextures(texture) : null
+  const reliefTextures = createBoardReliefTextures(texture)
   const surfaceMaterial = REALISTIC_BOARD_SURFACE_MATERIAL
 
-  const material =
-    renderingMode === "realistic"
-      ? new THREE.MeshPhysicalMaterial({
-          ...sharedMaterialOptions,
-          bumpMap: reliefTextures?.bumpMap ?? null,
-          bumpScale: surfaceMaterial.bumpScale,
-          normalMap: reliefTextures?.normalMap ?? null,
-          normalScale: new THREE.Vector2(
-            surfaceMaterial.normalScale,
-            surfaceMaterial.normalScale,
-          ),
-          roughnessMap: reliefTextures?.roughnessMap ?? null,
-          // The relief maps carry the material values per pixel: matte solder
-          // mask and ENIG-finished exposed copper need different responses.
-          roughness: 1,
-          metalnessMap: reliefTextures?.metalnessMap ?? null,
-          metalness: 1,
-          clearcoat: surfaceMaterial.clearcoat,
-          clearcoatRoughness: surfaceMaterial.clearcoatRoughness,
-          envMapIntensity: 0.85,
-        })
-      : new THREE.MeshBasicMaterial(sharedMaterialOptions)
+  const material = new THREE.MeshPhysicalMaterial({
+    ...sharedMaterialOptions,
+    bumpMap: reliefTextures?.bumpMap ?? null,
+    bumpScale: surfaceMaterial.bumpScale,
+    normalMap: reliefTextures?.normalMap ?? null,
+    normalScale: new THREE.Vector2(
+      surfaceMaterial.normalScale,
+      surfaceMaterial.normalScale,
+    ),
+    roughnessMap: reliefTextures?.roughnessMap ?? null,
+    // The relief maps carry the material values per pixel: matte solder mask
+    // and ENIG-finished exposed copper need different responses.
+    roughness: 1,
+    metalnessMap: reliefTextures?.metalnessMap ?? null,
+    metalness: 1,
+    clearcoat: surfaceMaterial.clearcoat,
+    clearcoatRoughness: surfaceMaterial.clearcoatRoughness,
+    envMapIntensity: 0.85,
+  })
   const mesh = new THREE.Mesh(planeGeom, material)
   mesh.position.set(
     boardOutlineBounds.centerX,
@@ -103,7 +96,6 @@ export function createTextureMeshes(
   isFaux: boolean = false,
   options: {
     shadowsEnabled?: boolean
-    renderingMode?: RenderingMode
   } = {},
 ): THREE.Mesh[] {
   const meshes: THREE.Mesh[] = []
@@ -119,7 +111,6 @@ export function createTextureMeshes(
       usePolygonOffset: true,
       renderOrder: 1,
       isFaux,
-      renderingMode: options.renderingMode,
     },
     boardData,
   )
@@ -142,7 +133,6 @@ export function createTextureMeshes(
       usePolygonOffset: true,
       renderOrder: 1,
       isFaux,
-      renderingMode: options.renderingMode,
     },
     boardData,
   )
